+++ docs/conf.py
@@ -293,4 +293,7 @@ texinfo_documents = [
 
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'python': ('https://docs.python.org/3.5', None)}
+intersphinx_mapping = {
+    'argh': ('http://argh.readthedocs.org/en/latest/', None),
+    'python': ('https://docs.python.org/3.5', None),
+}