+++ CHANGES.rst
@@ -3,6 +3,8 @@ Version 1.1.0
 
 Release TDB
 
+- Add ``henson.cli.register_commands`` to extend the command line interface
+
 Version 1.0.0
 -------------